Adding a mechanism to create a proxy widget for several proxies.
Adding a new class, pqProxiesWidget, which can thought of as a
pqProxyWidget for several proxies. It creates pqProxyWidget instances
for each proxy and manages that. It is designed to be used in custom
applications that need to customize the Properties panels to show
properties from a collection of proxies rather than just one at a time.

Change-Id: I96b22ca7957fe0e2eeef981359dba4dc30ff213d